@@ -17,9 +17,9 @@
     "nav-frontend-alertstriper-style": "0.2.33",
     "nav-frontend-chevron": "1.0.2",
     "nav-frontend-chevron-style": "0.3.4",
-    "nav-frontend-core": "4.0.5",
-    "nav-frontend-grid": "1.0.14",
-    "nav-frontend-grid-style": "0.2.15",
+    "nav-frontend-core": "^4.0.5",
+    "nav-frontend-grid": "^1.0.14",
+    "nav-frontend-grid-style": "^0.2.15",
     "nav-frontend-ikoner-assets": "^0.2.27",
     "nav-frontend-js-utils": "^1.0.3",
     "nav-frontend-knapper": "1.0.5",
@@ -37,9 +37,9 @@
     "nav-frontend-spinner-style": "0.2.4",
     "nav-frontend-typografi": "2.0.3",
     "nav-frontend-typografi-style": "1.0.8",
-    "prop-types": "^15.6.0",
+    "prop-types": "^15.6.1",
     "query-string": "5.1.0",
-    "react": "^16.2.0",
+    "react": "^16.4.0",
     "react-day-picker": "7.1.6",
     "react-dom": "16.3.2",
     "react-intl": "2.4.0",
