@@ -10,7 +10,9 @@ import { AppState } from '../../reducer';
 import { hentFornavn } from '../../utils/utils';
 import { SKJEMA_PATH } from '../../utils/konstanter';
 import LenkeAvbryt from '../../komponenter/knapper/lenke-avbryt';
-const veilederSvg = require('../../ikoner/illustrasjon-start.svg');
+import Knappervertikalt from '../../komponenter/knapper/knapper-vertikalt';
+
+const personSvg = require('./person-komprimert.svg');
 
 interface StateProps {
     innloggingsInfo: InnloggingsInfoState;
@@ -24,23 +26,38 @@ export class Start extends React.Component<RouteComponentProps<MatchProps> & Sta
         const {name} = innloggingsInfo.data;
         return (
             <section className="startside">
-                <div className="figur__wrapper">
-                    <img src={veilederSvg} className="figur"/>
+                <div className="startside__banner">
+                    <div className="limit">
+                        <div className="startside__intro">
+                            <p className="typo-undertittel">
+                                <FormattedMessage id="overskrift-start-dialog" values={{fornavn: hentFornavn(name)}}/>
+                            </p>
+                            <p className="typo-normal">
+                                <FormattedMessage id="beskrivelse-start-dialog"/>
+                            </p>
+                        </div>
+                        <img
+                            className="startside__intro__ikon"
+                            src={personSvg}
+                            alt="Person-ikon"
+                        />
+                    </div>
                 </div>
                 <div className="innhold-container">
                     <div className="innhold">
                         <Innholdstittel className="tittel">
-                            <FormattedMessage id="overskrift-start" values={{fornavn: hentFornavn(name)}}/>
+                            <FormattedMessage id="overskrift-start"/>
                         </Innholdstittel>
                         <Normaltekst className="beskrivelse"><FormattedMessage id="beskrivelse-start"/></Normaltekst>
-                        <KnappBase
-                            type="hoved"
-                            className="knapp-start"
-                            onClick={() => history.push(`${SKJEMA_PATH}/0`)}
-                        >
-                            <FormattedMessage id="knapp-start"/>
-                        </KnappBase>
-                        <LenkeAvbryt />
+                        <Knappervertikalt>
+                            <KnappBase
+                                type="hoved"
+                                onClick={() => history.push(`${SKJEMA_PATH}/0`)}
+                            >
+                                <FormattedMessage id="knapp-start"/>
+                            </KnappBase>
+                            <LenkeAvbryt tekstId="avbryt-lenke"/>
+                        </Knappervertikalt>
                     </div>
                 </div>
             </section>
