@@ -7,12 +7,13 @@ import { Element } from 'nav-frontend-typografi';
 
 interface LenkeAvbrytProps {
     classname?: string;
+    tekstId?: string;
 }
 
-function LenkeAvbryt({classname}: LenkeAvbrytProps) {
+function LenkeAvbryt({classname, tekstId = 'avbryt-lenke-registrering'}: LenkeAvbrytProps) {
     return(
         <Link className={classnames('lenke-avbryt', classname)} to={AVBRYT_PATH}>
-            <Element tag="span"><FormattedMessage  id="avbryt-lenke"/></Element>
+            <Element tag="span"><FormattedMessage  id={tekstId}/></Element>
         </Link>
     );
 }
