@@ -7,11 +7,11 @@ import * as Adapter from 'enzyme-adapter-react-16';
 import SjekkRegistreringstatus from './sjekk-registreringstatus';
 import {
     dispatchRegistreringstatus,
-    promiseWithSetTimeout, shallowWithStore, mountWithStore
+    promiseWithSetTimeout, mountWithStoreAndIntl, shallowwithStoreAndIntl
 } from '../../test/test-utils';
 import SblRegistrering from '../../sider/oppsummering/sbl-registrering';
 import { create } from '../../store';
-import { sendBrukerTilVeientilarbeid } from './utils';
+import AlleredeRegistrert from '../../sider/allerede-registrert/allerede-registrert';
 
 enzyme.configure({adapter: new Adapter()});
 
@@ -28,25 +28,20 @@ describe('<SjekkRegistreringstatus />', () => {
         const store = create();
         dispatchRegistreringstatus({underOppfolging: false, oppfyllerKrav: false}, store);
 
-        const wrapper = shallowWithStore(<SjekkRegistreringstatus />, store);
+        const wrapper = shallowwithStoreAndIntl(<SjekkRegistreringstatus />, store);
 
         expect(wrapper.find(SblRegistrering)).to.have.length(1);
 
     });
 
-    it('skal sende bruker til veien til arbeid om den er under oppfølging', () => {
+    it('skal sende bruker til AlleredeRegistrert om den er under oppfølging', () => {
         const store = create();
-
-        const sendBrukerTilVeientilarbeidSpy = sandbox.spy(sendBrukerTilVeientilarbeid);
-        const config = {
-            sendBrukerTilVeientilarbeid: sendBrukerTilVeientilarbeidSpy,
-        };
 
         dispatchRegistreringstatus({underOppfolging: true, oppfyllerKrav: false}, store);
 
-        mountWithStore(<SjekkRegistreringstatus config={config}/>, store);
+        const wrapper = mountWithStoreAndIntl(<SjekkRegistreringstatus/>, store);
 
-        expect(sendBrukerTilVeientilarbeidSpy.called).to.be.equal(true);
+        expect(wrapper.find(AlleredeRegistrert)).to.have.length(1);
     });
     it('Skal rendre innhold dersom bruker oppfyller krav og ikke er under oppfølging', () => {
         const store = create();
@@ -59,7 +54,7 @@ describe('<SjekkRegistreringstatus />', () => {
             </SjekkRegistreringstatus>
         );
 
-        const wrapper = mountWithStore(component, store);
+        const wrapper = mountWithStoreAndIntl(component, store);
 
         return promiseWithSetTimeout()
             .then(() => {
